+++ src/main/java/org/cookieandkakao/babting/common/properties/KakaoClientProperties.java
@@ -1,0 +1,13 @@
+package org.cookieandkakao.babting.common.properties;
+
+import java.util.Set;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "oauth2.client.kakao")
+public record KakaoClientProperties(
+    String clientId,
+    String clientSecret,
+    String redirectUri,
+    Set<String> scope) {
+
+}
