--- src/main/java/org/cookieandkakao/babting/domain/member/properties/KakaoClientProperties.java
@@ -1,10 +1,0 @@
-package org.cookieandkakao.babting.domain.member.properties;
-
-import java.util.Set;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "oauth2.client.kakao")
-public record KakaoClientProperties(String clientId, String clientSecret, String redirectUri,
-                                    Set<String> scope) {
-
-}
