--- src/main/java/org/cookieandkakao/babting/domain/member/properties/KakaoProviderProperties.java
@@ -1,9 +1,0 @@
-package org.cookieandkakao.babting.domain.member.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "oauth2.provider.kakao")
-public record KakaoProviderProperties(String authorizationUri, String tokenUri,
-                                      String userInfoUri) {
-
-}
